requirements.txt
boto3
python-dotenv
pymongo
pandas
openpyxl
